requirements.txt
clipboard
pygame-ce
setuptools
pytest
pytest-coverage
pillow
